requirements.txt
openai
PyYAML
pydantic
python-dotenv
pybars3
rich
